Split test_coordinates into different files
Also moves it into a separate folder to keep everything tidy.
