+++ testsuite/MDAnalysisTests/coordinates/test_coordinates.py
@@ -1,0 +1,231 @@
+# -*- Mode: python; tab-width: 4; indent-tabs-mode:nil; coding:utf-8 -*-
+# vim: tabstop=4 expandtab shiftwidth=4 softtabstop=4 fileencoding=utf-8
+#
+# MDAnalysis --- http://www.MDAnalysis.org
+# Copyright (c) 2006-2015 Naveen Michaud-Agrawal, Elizabeth J. Denning, Oliver
+# Beckstein and contributors (see AUTHORS for the full list)
+#
+# Released under the GNU Public Licence, v2 or any higher version
+#
+# Please cite your use of MDAnalysis in published work:
+#
+# N. Michaud-Agrawal, E. J. Denning, T. B. Woolf, and O. Beckstein.
+# MDAnalysis: A Toolkit for the Analysis of Molecular Dynamics Simulations.
+# J. Comput. Chem. 32 (2011), 2319--2327, doi:10.1002/jcc.21787
+#
+
+import MDAnalysis
+import MDAnalysis as mda
+import MDAnalysis.coordinates
+import MDAnalysis.coordinates.core
+from MDAnalysis import NoDataError
+
+import numpy as np
+from numpy.testing import *
+from nose.plugins.attrib import attr
+
+from MDAnalysisTests.datafiles import (
+    PSF, DCD, DCD_empty, PDB_small, XPDB_small, PDB_closed, PDB_multiframe,
+    PDB_full, PDB, CRD, XTC, TRR, GRO, DMS, CONECT, INC_PDB, XYZ, XYZ_bz2,
+    XYZ_psf, PRM, TRJ, TRJ_bz2, PRMpbc, TRJpbc_bz2, PRMncdf, NCDF, PQR,
+    PDB_sub_dry, TRR_sub_sol, PDB_sub_sol, TRZ, TRZ_psf, LAMMPSdata,
+    LAMMPSdata_mini, PSF_TRICLINIC, DCD_TRICLINIC, PSF_NAMD_TRICLINIC,
+    DCD_NAMD_TRICLINIC, GMS_ASYMOPT, GMS_SYMOPT, GMS_ASYMSURF, XYZ_mini,
+    PFncdf_Top, PFncdf_Trj, INPCRD, XYZ_five, two_water_gro, DLP_CONFIG,
+    DLP_CONFIG_order, DLP_CONFIG_minimal, DLP_HISTORY, DLP_HISTORY_order,
+    DLP_HISTORY_minimal, PDB_xlserial)
+
+from MDAnalysisTests.plugins.knownfailure import knownfailure
+
+import os
+import errno
+import tempfile
+import itertools
+
+
+def atom_distance(a, b):
+    """Calculate the distance between two atoms a and b."""
+    r = a.position - b.position
+    return np.sqrt(np.sum(r ** 2))
+
+
+
+
+class TestINPCRDReader(TestCase):
+    """Test reading Amber restart coordinate files"""
+    def _check_ts(self, ts):
+        # Check a ts has the right values in
+        ref_pos = np.array([[6.6528795, 6.6711416, -8.5963255],
+                        [7.3133773, 5.8359736, -8.8294175],
+                        [8.3254058, 6.2227613, -8.7098593],
+                        [7.0833200, 5.5038197, -9.8417650],
+                        [7.1129439, 4.6170351, -7.9729560]])
+        for ref, val in itertools.izip(ref_pos, ts._pos):
+            assert_allclose(ref, val)
+
+    def test_reader(self):
+        from MDAnalysis.coordinates.INPCRD import INPReader
+
+        r = INPReader(INPCRD)
+
+        assert_equal(r.n_atoms, 5)
+        self._check_ts(r.ts)
+
+    def test_universe_inpcrd(self):
+        u = MDAnalysis.Universe(XYZ_five, INPCRD)
+
+        self._check_ts(u.trajectory.ts)
+
+    def test_universe_restrt(self):
+        u = MDAnalysis.Universe(XYZ_five, INPCRD, format='RESTRT')
+        self._check_ts(u.trajectory.ts)
+
+
+
+
+
+
+
+### Does not implement Reader.remarks, Reader.header, Reader.title, Reader.compounds
+### because the PDB header data in trajectory.metadata are already parsed; should perhaps
+### update the PrimitivePDBReader to do the same. [orbeckst]
+#class TestPDBReader_Metadata(_PDBMetadata):
+#    permissive = False
+
+class TestPSF_CRDReader(_SingleFrameReader):
+    def setUp(self):
+        self.universe = mda.Universe(PSF, CRD)
+        self.prec = 5  # precision in CRD (at least we are writing %9.5f)
+
+
+class TestPSF_PDBReader(TestPDBReader):
+    def setUp(self):
+        # mda.core.flags['permissive_pdb_reader'] = False
+        self.universe = mda.Universe(PSF, PDB_small, permissive=False)
+        self.prec = 3  # 3 decimals in PDB spec http://www.wwpdb.org/documentation/format32/sect9.html#ATOM
+
+    def test_uses_Biopython(self):
+        from MDAnalysis.coordinates.PDB import PDBReader
+
+        assert_(isinstance(self.universe.trajectory, PDBReader), "failed to choose Biopython PDBReader")
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+class TestChainReader(TestCase):
+    def setUp(self):
+        self.universe = mda.Universe(PSF, [DCD, CRD, DCD, CRD, DCD, CRD, CRD])
+        self.trajectory = self.universe.trajectory
+        self.prec = 3
+        # dummy output DCD file
+        fd, self.outfile = tempfile.mkstemp(suffix=".dcd")
+        os.close(fd)
+
+    def tearDown(self):
+        try:
+            os.unlink(self.outfile)
+        except:
+            pass
+        del self.universe
+
+    def test_next_trajectory(self):
+        self.trajectory.rewind()
+        self.trajectory.next()
+        assert_equal(self.trajectory.ts.frame, 1, "loading frame 2")
+
+    def test_n_atoms(self):
+        assert_equal(self.universe.trajectory.n_atoms, 3341, "wrong number of atoms")
+
+    def test_n_frames(self):
+        assert_equal(self.universe.trajectory.n_frames, 3 * 98 + 4, "wrong number of frames in chained dcd")
+
+    def test_iteration(self):
+        for ts in self.trajectory:
+            pass  # just forward to last frame
+        assert_equal(self.trajectory.n_frames - 1, ts.frame,
+                     "iteration yielded wrong number of frames (%d), should be %d"
+                     % (ts.frame, self.trajectory.n_frames))
+
+    def test_jump_lastframe_trajectory(self):
+        self.trajectory[-1]
+        #print self.trajectory.ts, self.trajectory.ts.frame
+        assert_equal(self.trajectory.ts.frame + 1, self.trajectory.n_frames, "indexing last frame with trajectory[-1]")
+
+    def test_slice_trajectory(self):
+        frames = [ts.frame for ts in self.trajectory[5:17:3]]
+        assert_equal(frames, [5, 8, 11, 14], "slicing dcd [5:17:3]")
+
+    def test_full_slice(self):
+        trj_iter = self.universe.trajectory[:]
+        frames = [ts.frame for ts in trj_iter]
+        assert_equal(frames, np.arange(self.universe.trajectory.n_frames))
+
+    def test_frame_numbering(self):
+        self.trajectory[98]  # index is 0-based and frames are 0-based
+        assert_equal(self.universe.trajectory.frame, 98, "wrong frame number")
+
+    def test_frame(self):
+        self.trajectory[0]
+        coord0 = self.universe.atoms.coordinates().copy()
+        # forward to frame where we repeat original dcd again:
+        # dcd:0..97 crd:98 dcd:99..196
+        self.trajectory[99]
+        assert_array_equal(self.universe.atoms.coordinates(), coord0,
+                           "coordinates at frame 1 and 100 should be the same!")
+
+    @knownfailure("time attribute not implemented for chained reader", ValueError)
+    def test_time(self):
+        self.trajectory[30]  # index is 0-based but frames are 1-based
+        assert_almost_equal(self.universe.trajectory.time, 31.0, 5,
+                            err_msg="wrong time of frame")
+
+    @dec.slow
+    def test_write_dcd(self):
+        """test that ChainReader written dcd (containing crds) is correct (Issue 81)"""
+        W = MDAnalysis.Writer(self.outfile, self.universe.atoms.n_atoms)
+        for ts in self.universe.trajectory:
+            W.write(self.universe)
+        W.close()
+        self.universe.trajectory.rewind()
+        u = MDAnalysis.Universe(PSF, self.outfile)
+        for (ts_orig, ts_new) in itertools.izip(self.universe.trajectory, u.trajectory):
+            assert_almost_equal(ts_orig._pos, ts_new._pos, self.prec,
+                                err_msg="Coordinates disagree at frame %d" % ts_orig.frame)
+
+
+class TestChainReaderFormats(TestCase):
+    """Test of ChainReader with explicit formats (Issue 76)."""
+
+    @attr('issue')
+    def test_set_all_format_tuples(self):
+        universe = MDAnalysis.Universe(GRO, [(PDB, 'pdb'), (XTC, 'xtc'), (TRR, 'trr')])
+        assert_equal(universe.trajectory.n_frames, 21)
+
+    @attr('issue')
+    def test_set_one_format_tuple(self):
+        universe = MDAnalysis.Universe(PSF, [(PDB_small, 'pdb'), DCD])
+        assert_equal(universe.trajectory.n_frames, 99)
+
+    @attr('issue')
+    def test_set_all_formats(self):
+        universe = MDAnalysis.Universe(PSF, [PDB_small, PDB_closed], format='pdb')
+        assert_equal(universe.trajectory.n_frames, 2)
